Did a spike of a MySQL driver for Self.
I just wanted to see how hard it would be to make Self usable for simple web applications.

The driver doesn't do much yet, but the basics for speaking the MySQL socket protocol are mostly there, and it can do queries and sometimes even convert the data to/from objects.

There's an implementation of SHA1 in there, not well tested.

Also added a few miscellaneous methods to some core objects (like byteVectors).

Oh, and there are some rudimentary readStream and writeStream objects. (I just pulled the reading and writing methods from traits abstractFile up into a couple of mixins, one for reading and one for writing. So now you can create a writeStream, which has the same writing methods that a file has, but which is backed by a string.)