Friend or Foe? Cope for differences in gcc/clang
Gcc and clang are handing 'friend' functions differently
when they are defined within a class but declared outside.
Additionally, gcc's interpretation of this changed a few
years ago (-> -ffriend-injection).

In order to make self compile with both,
a) true 'friend' functions have been left mostly as is,
   probably with its definition moved into in a .cpp file;
b) 'friends' that are such just as to allow a shorthand invocation
   are now just 'static' and hence have to be fully qualified
   (eg, findNMethod(void*) => nmethod::findNMethod(void*)); and
c) 'friends' that are actually special constructors have been made
   into static functions, as well, but are also given a top-level
   namespace name via a static inline (eg, as_memOop(...)).

For a few other functions, way c) is also taken.