Rework the Mac OS X starter for the Self VM.
This change makes Self on OS X a proper application,
double-clickable, with an optional redirection of the
stdin/stdout to a builtin console.
  This remains Experimental. There are spontaneous, strage
issues with core graphics, that probably relate to carbon windows
and threading. Also, when starting with command line options,
enabling the Console results in the desktop constantly stopping,
you have to manually 'desktop go' it.
  Apart from that, everything looks fine.

Conflicts:

	vm/src/CMakeLists.txt