asm: Add NetBSD to the gas(1) #ifdef branches
This is not right, as NetBSD can also be built with clang as the
default compiler, but I don't have such configuration handy at the
moment to test.

In general these files really just choose between gas and clang syntax
and the choice should be made once instead of repeating it in a
slightly different way each time, several times in each file.