allocation.cpp: Add NetBSD
Implement true_size_of_malloced_obj() by just asking jemalloc
directly, it has an API for that, though it's in a separate
library (that just provide public names for wrappers that forward to
the real thing hidden back in libc).

Caveat: the new jemalloc that has this API is only available
in NetBSD 9 or later.