sparc: add # ifdef INTERFACE_PRAGMAS
6aa20c4 disabled #pragma interface for gcc with a comment that it
"fails with interface pragmas".  It does, but b/c it's the build
process that does it wrong.  The current build process includes the
precompiled header via command line -include, so all #pragma interface
from the inline headers are processed _before_ the source file has a
chance to say it's a #pragma implementation for them.

unless #pragma implementation is in effect.  Since it is not in effect
in any file, the code for the inline methods is never emitted and
there are tons of undefined symbols at the link step if you compile
with -O0 and the inline functions are not actually inlined.

Most of the inline functions are only declared inline in the class
body and their actual definitions are provided in the "inline"
headers.  That kind of inline is treated by gcc as just plain inline
and is not inlined at -O0.  Either the definitions needs to be in the
class body (which goes against the foo.hh vs foo_inline.hh split to
begin with) or always_inline attribute needs to be specified for them.