Fix precompiled header vs. #pragma implementation
-include on the command line creates a problem: headers from the
_precompiled.hh are processed before a source file has a chance to
state its #pragma implementation, hence the kludge with the
INTERFACE_PRAGMAS ifdefs in the sources.

The combined effect of not specifying the pragmas and the
-fkeep-inline-functions flag is *huge* object files with hundreds of
kilobytes of the same object code emitted in each of them.

This commit changes makeDeps to includeGIInEachIncl so that pragma's
DTRT.  The ifdefs should probably be disposed of, but for now keep
them and supply the -D instead as a quick, small, low risk change.