setup.cmake: get rid of -fkeep-inline-functions
Fixing pragmas vs. the precompiled headers helps a lot, but there's
still enough inline stuff floating around to make the object files
full of object code from system headers and core classes - including
static inlines that end up in the final executable multiple times b/c
they are static and cannot be merged.

Not why this flag was ever enabled.  I suspect it was a workaround for
the pragma problem (that causes inline functions to be emitted in the
"implementation" object file).